+++ src/main.h
@@ -19,7 +19,7 @@
 #pragma once
 
 #ifdef BUILD_REL
-    #define MOD_VERSION "Thinker Mod v5.1 - The Will to Power mod v393"
+    #define MOD_VERSION "Thinker Mod v5.1 - The Will to Power mod v340"
 #else
     #define MOD_VERSION "Thinker Mod develop build"
 #endif
@@ -446,6 +446,9 @@ struct Config {
     int promotion_probability_denominator = 1;
     double lab_mineral_worth = 0.5;
     bool road_magtube_fixed_movement_rate = false;
+    int right_of_passage_agreement = 0;
+    bool scorched_earth = false;
+    bool destroy_captured_base_defense = false;
     // AI configurations
     bool ai_useWTPAlgorithms;
     bool wtp_enabled_factions[MaxPlayerNum];